feat(modularization): Begin Phase 3 — introduce orchestration engine
- Add src/orchestration/engine.py with act_on_search orchestration
- Add src/orchestration/__init__.py re-export
- Update src/actions.py to delegate act_on_search to orchestration engine (shim)
- All tests still pass (535), coverage 97.82%

Note: Keep actions.py as compatibility shim; main/scheduler unchanged for now.
